pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-studio-poc"
version = "0.1.0"
description = "Persistent multi-agent AI studio proof of concept"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "autogen-agentchat",
    "autogen-ext[openai]",
    "click",
    "GitPython",
    "openai",
    "openai-agents>=0.12.4,<1",
    "pydantic>=2",
    "python-dotenv",
    "SQLAlchemy",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["agents", "skills", "sessions"]
